Update to examples directory
