+++ examples/chatbot/modeling.py
@@ -10,7 +10,7 @@ from typing import Literal
 
 import datasets
 
-from tasks.chatbot import config as chatbot_config
+from examples.chatbot import config as chatbot_config
 from zeno_build.cache_utils import CacheLock, fail_cache, get_cache_path
 from zeno_build.models.chat_generate import generate_from_chat_prompt
 from zeno_build.prompts.chat_prompt import ChatMessages, ChatTurn
