+++ examples/chatbot/report/images/example-gpt3.5-correct-1.png
