+++ examples/chatbot/report/images/example-gpt3.5-hallucination-2.png
