+++ examples/chatbot/report/images/example-gpt3.5-probing-1.png
