+++ examples/chatbot/report/images/example-gpt3.5-repeat-1.png
