+++ examples/chatbot/report/images/model-bylength-chrf.png
