+++ examples/chatbot/report/images/model-overall-bertscore.png
