+++ examples/chatbot/report/images/model-overall-chrf.png
