+++ examples/chatbot/report/images/model-overall-coherence.png
