+++ examples/chatbot/report/images/model-overall-lengthratio.png
