+++ examples/chatbot/report/images/prompt-bycontext-bertscore.png
