+++ examples/chatbot/report/images/prompt-overall-bertscore.png
