+++ examples/chatbot/report/images/prompt-overall-chrf.png
