+++ examples/chatbot/report/images/window-bycontext-chrf.png
