+++ examples/chatbot/report/images/window-bylength-chrf.png
