+++ examples/chatbot/report/images/window-overall-chrf.png
