+++ examples/summarization/__init__.py
@@ -1,0 +1,1 @@
+"""A package for analyzing text summarization."""
