+++ examples/summarization/modeling.py
@@ -9,7 +9,7 @@ import os
 
 import datasets
 
-from tasks.summarization import config as summarization_config
+from examples.summarization import config as summarization_config
 from zeno_build.cache_utils import get_cache_path
 from zeno_build.models.text_generate import generate_from_text_prompt
