+++ pyproject.toml
@@ -26,7 +26,7 @@ test = ["mypy>=1.2.0", "pre-commit>=3.2.0", "pytest>=6.0.0", "black>=23.3.0"]
 
 [tool.hatch.build]
 include = ["*.py"]
-exclude = ["*_test.py", "tasks/**"]
+exclude = ["*_test.py", "tasks/**", "examples/**"]
 only-packages = true
 
 [tool.hatch.build.targets.wheel]
