--- tasks/summarization/__init__.py
@@ -1,1 +1,0 @@
-"""A package for analyzing text summarization tasks."""
